# apps/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go + Python 퀀트 플랫폼 API 서버입니다.",
        "title": "Quant Trading API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/paper/balance": {
            "get": {
                "description": "KIS Open API로 잔고와 보유종목을 조회합니다. 모의/실전 분기는 엔진의 KIS_MODE 환경변수가 결정합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paper"
                ],
                "summary": "모의투자 계좌 잔고 조회 (KIS API 엔진 프록시)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/paper/orders": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paper"
                ],
                "summary": "모의투자 주문/체결 내역 (KIS API 엔진 프록시)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "YYYYMMDD — 생략 시 오늘",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "YYYYMMDD — 생략 시 오늘",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "{ symbol, qty, side: buy|sell, order_type: market|limit, price? }",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paper"
                ],
                "summary": "모의투자 주문 (KIS API 엔진 프록시)",
                "parameters": [
                    {
                        "description": "주문 payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/paper/quote/{symbol}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paper"
                ],
                "summary": "국내주식 현재가 (KIS API 엔진 프록시)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "6자리 숫자 코드 또는 .KS/.KQ suffix 포함",
                        "name": "symbol",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}